requirements.txt
black==23.12.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dacite==1.8.1
databind==4.5.1
databind.core==4.5.1
databind.json==4.5.1
Deprecated==1.2.14
docspec==2.2.1
docspec-python==2.2.1
docstring-parser==0.11
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
iota-sdk @ file:///home/pi/iota-sdk/bindings/python
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nr-date==2.1.0
nr-stream==1.1.5
nr.util==0.8.12
packaging==24.0
paho-mqtt==2.0.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
pydoc-markdown==4.8.2
pyhumps==3.8.0
pytest==8.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.10.0
setuptools-rust==1.9.0
tomli==2.0.1
tomli_w==1.0.0
typeapi==2.2.1
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
wrapt==1.16.0
yapf==0.40.2
zipp==3.18.1
